refactor: centralize task manager
