+++ agents/task_management/subgoal_generator.py
@@ -1,0 +1,3 @@
+class SubGoalGenerator:
+    async def generate_subgoals(self, description: str, context: dict) -> list[str]:
+        return [description]
